refactor(#546, #547, #548): extract GameControllerTrait, GameStatsCalculator, GameDifficulty, and use GateInterface
#546 — GameControllerTrait with shared helpers
- jsonBody(), json(), cleanDefinition(), loadSessionByToken() in one trait
- loadSessionByToken() now uses framework loadByKey() (one-liner)
- Both controllers use the trait, removing all duplicated methods

#547 — GameStatsCalculator and GameDifficulty
- GameStatsCalculator::build() replaces duplicate buildStats() in both controllers
- Accepts game type and configurable streak-breaker/win statuses
- GameDifficulty::dailyTier() replaces engine-specific implementations
- CrosswordEngine no longer depends on ShkodaEngine

#548 — GateInterface ownership check
- ShkodaController injects GateInterface via constructor DI
- Hand-rolled ownership check replaced with $gate->denies('update', $session, $account)
- Delegates to GameAccessPolicy::canModifySession() properly

Closes #546
Closes #547
Closes #548
